% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\docType{data}
\name{srhd_palette}
\alias{srhd_palette}
\title{SRHD Color Palette}
\format{
An object of class \code{character} of length 13.
}
\usage{
srhd_palette
}
\description{
A vector of hex color values used by SRHD.
}
\keyword{datasets}
